fix(tape): promote dword to size_t before adding header bytes
Gemini-flagged on PR #129: in cases 0x35 (Custom info) and the default
unknown-block branch, `block_size = d + 0x14 + 1` (and `+ 4 + 1`) does
the addition in uint32_t before assigning to size_t.  A malformed file
with `d` near UINT32_MAX wraps to a small block_size, slipping past
the subsequent `block_size > end - p` bounds check and desynchronising
the scanner.

Fix: cast `d` to size_t before the addition.  Other dword cases (0x11,
0x14, 0x15, 0x40) are unaffected — they mask to 24 bits which can't
overflow size_t.

+++ src/tape.cpp
@@ -686,7 +686,11 @@ void tape_scan_blocks()
         break;
       case 0x35: // Custom info
         if (!safe_read_dword(p, end, 0x11, d)) goto done;
-        block_size = d + 0x14 + 1;
+        // Promote to size_t BEFORE adding the header bytes; `d` is dword
+        // (uint32_t) and a malformed file with d near UINT32_MAX would
+        // wrap the addition in 32-bit, slipping past the bounds check
+        // below.  Same pattern applied to the default case.
+        block_size = static_cast<size_t>(d) + 0x14 + 1;
         break;
       case 0x40: // Snapshot
         if (!safe_read_dword(p, end, 0x02, d)) goto done;
@@ -697,7 +701,8 @@ void tape_scan_blocks()
         break;
       default: // Unknown block with 4-byte length
         if (!safe_read_dword(p, end, 0x01, d)) goto done;
-        block_size = d + 4 + 1;
+        // Promote before adding (see case 0x35 above).
+        block_size = static_cast<size_t>(d) + 4 + 1;
         break;
     }
